Here is a video of the Magic Leap AR version: https://vimeo.com/340217310
The frame rate is running at around 20 fps or more without screen recording. However it goes to below 10 fps for a few seconds when the simulation starts.
I have a few issues still that I am working on with the help of the team in Barendrecht
Sometimes the containers don't go all the way from the QC crane to the ground.
The view is a bit shaky.
Another issue that unfortunately I can't fix is that the Magic Leap forces a plane clipping of 37 cm. This means that you can't zoom in to any object closer than 37 cm.